+++ tests/test_inotify_buffer.py
@@ -119,7 +119,7 @@ def test_delete_watched_directory(p):
 
 
 @pytest.mark.timeout(5)
-@pytest.mark.skipIf("GITHUB_REF" not in os.environ)
+@pytest.mark.skipif("GITHUB_REF" not in os.environ, reason="sudo password prompt")
 def test_unmount_watched_directory_filesystem(p):
     mkdir(p("dir1"))
     mount_tmpfs(p("dir1"))